+++ src/Board/board/Rect.h
@@ -6,38 +6,6 @@
  * 
  * @brief  
  * @copyright
- * This source code is part of the Board project, a C++ library whose
- * purpose is to allow simple drawings in EPS, FIG or SVG files.
- * Copyright (C) 2007 Sebastien Fourey <http://www.greyc.ensicaen.fr/~seb/>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>. 
- * This source code is part of the Board project, a C++ library whose
- * purpose is to allow simple drawings in EPS, FIG or SVG files.
- * Copyright (C) 2007 Sebastien Fourey <http://www.greyc.ensicaen.fr/~seb/>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>. 
  */
 #ifndef _BOARD_RECT_H_
 #define _BOARD_RECT_H_
@@ -52,10 +20,10 @@ namespace LibBoard {
  * @brief Struct representing a rectangle on the plane. 
  */
 struct Rect {
-  float left;			/**< Coordinate of the left side. */
-  float top;			/**< Coordinate of the upper side. */
-  float width;			/**< Width of the rectangle. */
-  float height;		/**< Height of the rectangle. */
+  double left;			/**< Coordinate of the left side. */
+  double top;			/**< Coordinate of the upper side. */
+  double width;			/**< Width of the rectangle. */
+  double height;		/**< Height of the rectangle. */
   
   /** 
    * Rect constructor.
@@ -67,7 +35,7 @@ struct Rect {
    * 
    * @return 
    */
-  Rect( float left = 0.0f, float top = 0.0f, float width = 0.0f, float height = 0.0f )
+  Rect( double left = 0.0f, double top = 0.0f, double width = 0.0f, double height = 0.0f )
     :left( left ), top( top ), width( width ), height( height ) { } 
 
   Point topLeft() { return Point( left, top ); }
